refactor(backend): moved download to files router
