+++ new/src/backend/app/api/v1/routes/web/files.py
@@ -5,8 +5,9 @@ import traceback
 
 import pathlib
 
-from typing import Any
-from fastapi import Depends, Request, UploadFile, status
+from typing import Annotated, Any
+from fastapi import Depends, Path, Request, UploadFile, status
+from fastapi.responses import FileResponse
 from fastapi.routing import APIRouter
 from dependency_injector.wiring import inject, Provide
 
@@ -14,7 +15,7 @@ from api.v1.constants import ALLOWED_FILE_TYPES, MAX_SETUP_FILES_SIZE
 from api.v1.schemas.response import Response
 
 from core.dependency_injection.container import Container
-from core.exceptions.http import BadRequestError
+from core.exceptions.http import BadRequestError, NotFoundError
 
 
 from db.models.moleculeset_stats import AtomTypeCount, MoleculeSetStats
@@ -124,4 +125,31 @@ async def upload(
         raise BadRequestError(
             status_code=status.HTTP_400_BAD_REQUEST,
             detail="Error uploading files.",
+        ) from e
+
+
+@files_router.get("/{computation_id}/download")
+@inject
+async def download_charges(
+    request: Request,
+    computation_id: Annotated[str, Path(description="UUID of the computation.")],
+    io: IOService = Depends(Provide[Container.io_service]),
+) -> FileResponse:
+    """Returns a zip file with all charges for the provided computation."""
+
+    user_id = request.state.user.id if request.state.user is not None else None
+
+    try:
+        charges_path = io.get_charges_path(computation_id, user_id)
+        if not io.path_exists(charges_path):
+            raise FileNotFoundError()
+
+        archive_path = io.zip_charges(charges_path)
+
+        return FileResponse(path=archive_path, media_type="application/zip")
+    except FileNotFoundError as e:
+        raise NotFoundError(detail=f"Computation '{computation_id}' not found.") from e
+    except Exception as e:
+        raise BadRequestError(
+            status_code=status.HTTP_400_BAD_REQUEST, detail="Error downloading charges."
         ) from e
